fix: livechat rerender, feat: change it to hooks
