+++ src/components/layout/location-context.js
@@ -9,8 +9,6 @@ export const LocationProvider = ({
     children,
     setModalPayload,
     toggleModal,
-    is_livechat_interactive,
-    LC_API,
 }) => {
     return (
         <LocationContext.Provider
@@ -19,8 +17,6 @@ export const LocationProvider = ({
                 show_cookie_banner,
                 setModalPayload,
                 toggleModal,
-                is_livechat_interactive,
-                LC_API,
             }}
         >
             {children}
@@ -31,7 +27,6 @@ export const LocationProvider = ({
 LocationProvider.propTypes = {
     children: PropTypes.node.isRequired,
     has_mounted: PropTypes.bool,
-    is_livechat_interactive: PropTypes.bool,
     LC_API: PropTypes.object,
     setModalPayload: PropTypes.func,
     show_cookie_banner: PropTypes.bool,
